feat: Snuba single-query version of get_keys_and_top_values
Snuba tagstore now implements this as an actual query instead
of the superclass iterating over each key and making 2 queries for each.

This should allow the entire tag distribution sidebar to be rendered
with 1 API call.

One limitation is that we would have to do a second query to get the
count and values_seen, which may or may not be worth it as the client
should already know what these values are for the keys it is querying.